[VarExporter] Suppress deprecations for legacy fixtures
