+++ src/Symfony/Component/VarExporter/Tests/Fixtures/foo-serializable.php
@@ -2,7 +2,7 @@
 
 return \Symfony\Component\VarExporter\Internal\Hydrator::hydrate(
     $o = \Symfony\Component\VarExporter\Internal\Registry::unserialize([], [
-        'C:51:"Symfony\\Component\\VarExporter\\Tests\\FooSerializable":20:{a:1:{i:0;s:3:"bar";}}',
+        'C:60:"Symfony\\Component\\VarExporter\\Tests\\Fixtures\\FooSerializable":20:{a:1:{i:0;s:3:"bar";}}',
     ]),
     null,
     [],
